requirements.txt
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
conda==4.3.16
crashtest==0.3.1
cryptography==3.4.8
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==21.8.0
lockfile==0.12.2
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pipenv==2022.8.24
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.1.15
poetry-core==1.0.8
ptyprocess==0.7.0
pycosat==0.6.3
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
shellingham==1.4.0
SQLAlchemy==1.4.40
tomlkit==0.11.4
urllib3==1.26.12
virtualenv==20.16.4
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.8.1
